# docs/swagger.yaml
basePath: /
definitions:
...
  api.answerReportRequest:
    properties:
      id:
        type: integer
      response_text:
        type: string
    type: object
...
  api.createReportRequest:
    properties:
      request_text:
        type: string
      request_type:
        type: string
      reservation_id:
        type: integer
    type: object
...
  api.updateTicketByReportRequest:
    properties:
      reserevation_id:
        type: integer
      to_status_reservation:
        type: string
    type: object
...
  db.Report:
    properties:
      admin_id:
        type: integer
      id:
        type: integer
      request_text:
        type: string
      request_type:
        $ref: '#/definitions/db.RequestType'
      reservation_id:
        type: integer
      response_text:
        type: string
      user_id:
        type: integer
    type: object
  db.RequestType:
    enum:
    - PAYMENT-ISSUE
    - TRAVEL-DELAY
    - UNEXPECTED-RESERVED
    - ETC.
    type: string
    x-enum-varnames:
    - RequestTypePAYMENTISSUE
    - RequestTypeTRAVELDELAY
    - RequestTypeUNEXPECTEDRESERVED
    - RequestTypeETC
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for Bilitio ticket booking system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:3000/swagger/index.html
  title: Bilitio API
  version: "1.0"
paths:
...
  /reports:
    get:
      description: Get all reports from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Report'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all reports
      tags:
      - reports
    post:
      consumes:
      - application/json
      description: User creates a new report
      parameters:
      - description: Report creation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a report
      tags:
      - reports
  /reports/answer:
    post:
      consumes:
      - application/json
      description: Admin answers a user report
      parameters:
      - description: Report answer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.answerReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Answer a report
      tags:
      - reports
  /reports/ticket:
    put:
      consumes:
      - application/json
      description: Admin updates a ticket status based on a report
      parameters:
      - description: Ticket update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.updateTicketByReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update ticket by report
      tags:
      - reports
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"